Add documentation to `small_network::insights` and properly expose `unforgettable`
